Get a starting circle of rhombi

This function returns a data set of ten observations. Each observation is a triangle for one half of a rhombus. The location, direction, and shape of the rhombus is determined through the parameters.

return_starting_circle(point_x, point_y, rotation, side_length, shape)

Arguments

point_x
Starting x position for first point

point_y
Starting y position for first point

rotation
Rotation about the starting position in degrees

side_length
Length of one outside edge

shape
Ether "t" for thin rhombus or "T" for thick rhombus

Value

Data set of triangles that join to create a circle of rhombi

Details

The center point of the circle is given by the point_x and point_y parameters. This will appear in the data set as point1_x and point1_y. The rotation in degrees will turn the rhombi around this point while side_length will change the size. Finally, shape will provide either a thick or thin rhombus. A "t" for the shape parameter will yield a thin rhombus. Anything else will provide a thick one, but traditionally a "T" is used. Thin triangles all point into the circle (ten rhombi halfs) while thick triangles alternate pointing to each other (five full rhombi).

Examples

return_starting_circle(0, 0, 0, 1, "T")

#>    triangle shape point1_x point1_y  point2_x      point2_y  point3_x
#> 1         1     T        0        0  0.809017 -5.877853e-01  1.618034
#> 2         2     T        0        0  0.809017  5.877853e-01  1.618034
#> 3         3     T        0        0  0.809017  5.877853e-01  0.500000
#> 4         4     T        0        0 -0.309017  9.510565e-01  0.500000
#> 5         5     T        0        0 -0.309017  9.510565e-01 -1.309017
#> 6         6     T        0        0 -1.000000  1.224647e-16 -1.309017
#> 7         7     T        0        0 -1.000000  1.224647e-16 -1.309017
#> 8         8     T        0        0 -0.309017 -9.510565e-01 -1.309017
#> 9         9     T        0        0 -0.309017 -9.510565e-01  0.500000
#> 10       10     T        0        0  0.809017 -5.877853e-01  0.500000
#>      point3_y rhombus
#> 1   0.0000000       1
#> 2   0.0000000       1
#> 3   1.5388418       2
#> 4   1.5388418       2
#> 5   0.9510565       3
#> 6   0.9510565       3
#> 7  -0.9510565       4
#> 8  -0.9510565       4
#> 9  -1.5388418       5
#> 10 -1.5388418       5

return_starting_circle(5, 5, 180, 3, "t")

#>    triangle shape point1_x point1_y point2_x point2_y point3_x point3_y rhombus
#> 1         1     t 2.146830 5.927051        5        5 2.146830 4.072949       1
#> 2         2     t 3.236644 2.572949        5        5 2.146830 4.072949       2
#> 3         3     t 3.236644 2.572949        5        5 5.000000 2.000000       3
#> 4         4     t 6.763356 2.572949        5        5 5.000000 2.000000       4
#> 5         5     t 6.763356 2.572949        5        5 7.853170 4.072949       5
#> 6         6     t 7.853170 5.927051        5        5 7.853170 4.072949       6
#> 7         7     t 7.853170 5.927051        5        5 6.763356 7.427051       7
#> 8         8     t 5.000000 8.000000        5        5 6.763356 7.427051       8
#> 9         9     t 5.000000 8.000000        5        5 3.236644 7.427051       9
#> 10       10     t 2.146830 5.927051        5        5 3.236644 7.427051      10